rust: Move trainer bin to an example in the bot crate
