+++ example_bots_src/circler/Cargo.toml
@@ -11,5 +11,10 @@ strip = true
 lto = true
 
 [dependencies]
-wasmbot_client = { path = "../../libraries/Rust/wasmbot_client"}
+wasmbot_client = { path = "../../libraries/Rust/wasmbot_client" }
 
+[dev-dependencies]
+trainer = { path = "../../libraries/Rust/trainer" }
+
+[[example]]
+name = "trainer"
