+++ example_bots_src/circler/Makefile
@@ -1,14 +1,14 @@
 EXE_NAME = circler
 BOT_SRCS = src/lib.rs src/exploration.rs
+TRAINER_SRCS = examples/trainer.rs
 
 target/wasm32-unknown-unknown/release/$(EXE_NAME).wasm: $(BOT_SRCS)
 	cargo build --release --target=wasm32-unknown-unknown
 	chmod -x $@
 
-trainer: $(BOT_SRCS)
-	cargo build
-	BOT_LIBRARY_NAME=$(EXE_NAME) BOT_LIBRARY_PATH=../../../example_bots_src/circler/target/debug/ cargo build --manifest-path ../../libraries/Rust/trainer/Cargo.toml --target-dir ./target
-	ln -s ./target/debug/trainer ./$(EXE_NAME)-trainer
+trainer: $(BOT_SRCS) $(TRAINER_SRCS)
+	cargo build --example trainer
+	ln -s ./target/debug/examples/trainer ./$(EXE_NAME)-trainer
 
 .PHONY: wasm
 wasm: target/wasm32-unknown-unknown/release/$(EXE_NAME).wasm
