+++ example_bots_src/circler/examples/trainer.rs
@@ -1,0 +1,6 @@
+// This declaration is needed to include exported symbols in the binary
+extern crate circler;
+
+fn main() {
+	trainer::run();
+}
