+++ libraries/Rust/trainer/src/harness.rs
@@ -8,7 +8,12 @@ thread_local! {
 }
 
 #[no_mangle]
-pub extern "C" fn logFunction(log_level: i32, msg_ptr: usize, msg_len: usize) {
+extern "C" fn trainer_shutdown() {
+	panic!("Client shut down");
+}
+
+#[no_mangle]
+extern "C" fn trainer_logFunction(log_level: i32, msg_ptr: usize, msg_len: usize) {
 	let lifted = match lift_string(msg_ptr, msg_len) {
 		Ok(s) => s,
 		Err(_) => {
@@ -25,7 +30,7 @@ pub extern "C" fn logFunction(log_level: i32, msg_ptr: usize, msg_len: usize) {
 }
 
 #[no_mangle]
-pub extern "C" fn getRandomInt(min: i32, max: i32) -> i32 {
+extern "C" fn trainer_getRandomInt(min: i32, max: i32) -> i32 {
 	if max <= min {
 		return min;
 	}
