+++ libraries/Rust/wasmbot_client/src/platform.rs
@@ -1,0 +1,5 @@
+#[cfg(target_arch = "wasm32")]
+pub mod wasm;
+
+#[cfg(not(target_arch = "wasm32"))]
+pub mod trainer;
